# backend/cmd/server/docs/openapi.json
{
  "components": {
    "schemas": {
      "Error": {
        "properties": {
          "error": {
            "type": "string"
          }
        },
        "type": "object"
      },
      "LoginRequest": {
        "properties": {
          "password": {
            "example": "admin123",
            "format": "password",
            "type": "string"
          },
          "username": {
            "example": "admin",
            "type": "string"
          }
        },
        "required": [
          "username",
          "password"
        ],
        "type": "object"
      },
      "LoginResponse": {
        "properties": {
          "token": {
            "description": "JWT，用于后续请求头 Authorization: Bearer <token>",
            "type": "string"
          },
          "user": {
            "properties": {
              "id": {
                "type": "integer"
              },
              "role": {
                "enum": [
                  "admin",
                  "user"
                ],
                "type": "string"
              },
              "username": {
                "type": "string"
              }
            },
            "type": "object"
          }
        },
        "type": "object"
      }
    },
    "securitySchemes": {
      "BearerAuth": {
        "bearerFormat": "JWT",
        "description": "登录后返回的 JWT，在 Authorize 中填入 token 即可",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "告警路由与通知平台 API。需先调用 POST /api/v1/auth/login 获取 token，再在请求头中携带 Authorization: Bearer <token> 调用其他接口。",
    "title": "KK Alert API",
    "version": "1.0.0"
  },
  "openapi": "3.0.0",
  "paths": {
    "/api/v1/alerts": {
      "get": {
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "default": 1,
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "page_size",
            "schema": {
              "default": 20,
              "type": "integer"
            }
          },
          {
            "description": "告警 ID 模糊",
            "in": "query",
            "name": "alert_id",
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "标题模糊",
            "in": "query",
            "name": "title",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "datasource_id",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "severity",
            "schema": {
              "enum": [
                "critical",
                "warning",
                "info"
              ],
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "status",
            "schema": {
              "enum": [
                "firing",
                "resolved",
                "suppressed"
              ],
              "type": "string"
            }
          },
          {
            "description": "RFC3339",
            "in": "query",
            "name": "from",
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "RFC3339",
            "in": "query",
            "name": "to",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "items, total"
          }
        },
        "summary": "告警列表（分页、筛选）"
      }
    },
...
    "/api/v1/alerts/{id}": {
      "get": {
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "告警及通知记录"
          }
        },
        "summary": "告警详情"
      }
    },
    "/api/v1/auth/login": {
      "post": {
        "description": "成功后返回 token，在 Swagger 右上角「Authorize」中填入该 token 即可调用需认证的接口。",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            },
            "description": "成功"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "用户名或密码错误"
          }
        },
        "security": [],
        "summary": "登录获取 Token"
      }
    },
...
    "/api/v1/channels/{id}": {
      "delete": {
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "summary": "删除渠道"
      },
      "get": {
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "channel"
          }
        },
        "summary": "渠道详情"
      },
      "put": {
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "summary": "更新渠道"
      }
    },
    "/api/v1/channels/{id}/test": {
      "post": {
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "message"
          }
        },
        "summary": "测试渠道发送"
      }
    },
...
    "/api/v1/datasources/{id}": {
      "delete": {
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "summary": "删除数据源"
      },
      "get": {
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "datasource"
          }
        },
        "summary": "数据源详情"
      },
      "put": {
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "summary": "更新数据源"
      }
    },
    "/api/v1/datasources/{id}/test": {
      "post": {
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "message"
          }
        },
        "summary": "测试数据源连接"
      }
    },
    "/api/v1/reports/aggregate": {
      "get": {
        "parameters": [
          {
            "in": "query",
            "name": "group_by",
            "schema": {
              "enum": [
                "time",
                "datasource",
                "severity"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "聚合结果"
          }
        },
        "summary": "报表聚合"
      }
    },
    "/api/v1/reports/export": {
      "get": {
        "parameters": [
          {
            "in": "query",
            "name": "from",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "to",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "format",
            "schema": {
              "default": "json",
              "enum": [
                "json",
                "csv",
                "xlsx"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "文件流或 JSON"
          }
        },
        "summary": "导出报表"
      }
    },
    "/api/v1/reports/trend": {
      "get": {
        "parameters": [
          {
            "in": "query",
            "name": "hours",
            "schema": {
              "default": 24,
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "按小时计数"
          }
        },
        "summary": "趋势（近 N 小时）"
      }
    },
...
    "/api/v1/rules/batch": {
      "post": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "action": {
                    "enum": [
                      "enable",
                      "disable"
                    ],
                    "type": "string"
                  },
                  "ids": {
                    "items": {
                      "type": "integer"
                    },
                    "type": "array"
                  }
                },
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "summary": "批量启用/禁用"
      }
    },
    "/api/v1/rules/export": {
      "post": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "ids": {
                    "items": {
                      "type": "integer"
                    },
                    "type": "array"
                  }
                },
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "rules"
          }
        },
        "summary": "导出规则 JSON"
      }
    },
    "/api/v1/rules/import": {
      "post": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "mode": {
                    "enum": [
                      "add",
                      "overwrite"
                    ],
                    "type": "string"
                  },
                  "rules": {
                    "type": "array"
                  }
                },
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "imported, failed"
          }
        },
        "summary": "导入规则"
      }
    },
    "/api/v1/rules/test-match": {
      "post": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "matched, matched_alerts"
          }
        },
        "summary": "测试规则匹配"
      }
    },
    "/api/v1/rules/{id}": {
      "delete": {
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "summary": "删除规则"
      },
      "get": {
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "rule"
          }
        },
        "summary": "规则详情"
      },
      "put": {
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "summary": "更新规则"
      }
    },
    "/api/v1/settings": {
      "get": {
        "responses": {
          "200": {
            "description": "settings"
          }
        },
        "summary": "获取设置（如 retention_days）"
      },
      "put": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "retention_days": {
                    "type": "integer"
                  }
                },
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "summary": "更新设置（admin）"
      }
    },
...
    "/api/v1/templates/{id}": {
      "delete": {
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "summary": "删除模板"
      },
      "get": {
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "template"
          }
        },
        "summary": "模板详情"
      },
      "put": {
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "summary": "更新模板"
      }
    },
    "/api/v1/templates/{id}/preview": {
      "post": {
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "rendered"
          }
        },
        "summary": "预览模板渲染"
      }
    },
    "/api/v1/templates/{id}/set-default": {
      "put": {
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "summary": "设为默认模板"
      }
    },
    "/api/v1/users": {
      "get": {
        "responses": {
          "200": {
            "description": "list"
          }
        },
        "summary": "用户列表（admin）"
      },
      "post": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "created"
          }
        },
        "summary": "创建用户"
      }
    },
    "/api/v1/users/{id}": {
      "delete": {
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "summary": "删除用户"
      },
      "put": {
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "summary": "更新用户"
      }
    },
...
  },
  "security": [
    {
      "BearerAuth": []
    }
  ],
  "servers": [
    {
      "description": "Current host",
      "url": "/"
    }
  ]
}